Add allclose tolerance to policy equality
Previously, the policies needed to have values that don't differ by more than 1e-6 which is not possible when using Q-learning with certain learning rate and amount of data. This lead to teleport having 5 policies instead of 3 due to the estimated Q-values not being precise. The tolerance is env dependant.